requirements.txt
streamlit==1.32.0

langchain==0.1.14
langchain-community==0.0.31
langchain-groq==0.1.2

chromadb==0.4.24

sentence-transformers==2.6.1
transformers==4.38.2
huggingface-hub==0.21.4

torch>=2.0.0

pypdf
docx2txt
python-pptx
beautifulsoup4
requests
python-dotenv

crewai==0.28.8
crewai-tools==0.1.6